KeyChain: Implement key generation in Keystore
These are the KeyChain-side changes for implementing generation of
Keystore keys directly by KeyChain.

The core functionality is in this CL, where Keystore (via the standard
Android API) is requested to generate a key whose specification was
provided by the user via the DevicePolicyManager.

See the frameworks/base CL for more details.

Bug: 63388672
Test: cts-tradefed run commandAndExit cts-dev -a armeabi-v7a -m CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.DeviceOwnerTest#testKeyManagement -l DEBUG
Change-Id: I1f53cec0c31adef3b850282aaff11b655a54a760