KeyChain: Upgrade keys database
A new table was added to KeyChain's grants.db, storing the
user-visibilty state of each alias KeyChain manages.

Formerly it was not created during an upgrade, which led to existing
keys being classified as non-user-selectable and essentially not usable.

Now, during upgrade, the table for user-visibility is created and the
existing aliases are all set as user-selectable in it.

Test: m -j RunKeyChainRoboTests
Bug: 73898958
Change-Id: I3b92a957f4e949c13363769ece531af438895ff9