Remove the install option from the certificate selection prompt
Remove the ability to install a certificate via the certificate
selection prompt:
* If there are no keys available to choose from, do not show the dialog
at all.
* If there are keys to choose from, show the dialog but without the
install button.

Certificates can be installed by the user via Settings, so this change
does not remove any platform functionality.

While no data exists, anecdotal evidence suggests that certificate
installation is never done via this prompt.
On the other hand, data exists to show the certificate selection prompt
confuses users by standing in the way of a meaningful error message when
there are no certificates to choose from and is a source of support
calls.
In the context of TLS client certificates, where this dialog is most
commonly used, all other operating systems do not display a prompt to
the user when no client certificates are available.

Bug: 114199250
Test: Manual, connected to a TLS server which requires client authentication.
Change-Id: I1f1a7ab49f203941ed3f1ed1c31ae20519fda557