KeyChain: Remove screenlock dependency
KeyChain used to depend on KeyStore's initialization state since it
was importing keys using the FLAG_ENCRYPTED flag.

With that flag, imported keys were encrypted using an AES key derived
from the user's screen lock.
However, said AES key is stored in the filesystem next to the key
material and so does not provide additional security beyond filesystem
encryption.
Additionally, there are now better ways to protect key access via
Keymaster.

This change removes the use of this flag as well as the dependency on
KeyStore's initialization state.
That allows getting rid of the requirement for the device to have a
screen lock.

Test: atest CtsDevicePolicyManagerTestCases:com.android.cts.devicepolicy.MixedProfileOwnerTest#testKeyManagement
Bug: 120901345
Change-Id: Ief36a1fb4250c66c88d8d6aebbccbc43cd5ef2b0