KeyChain: Allow specifying UID of installed key
This is part of the work to unify manual and programmatic key
installation flows.

Since the manual key installation flow allowed installing keys to the
WiFi Keystore, this capability needs to be preserved and so to make
KeyChainService usable for installing WiFi keys, take the destination
UID as parameter to the installKeyPair method.

As a safety measure, only allow installation of keys to the "self" UID
and the WiFi UID (and the WiFi UID is only acceptable on the primary
user).

Test: Manual CtsVerifier tests: KeyChain Storage Test, CA Cert Notification Test
Test: cts-tradefed run commandAndExit cts-dev -m CtsDevicePolicyManagerTestCases -t  com.android.cts.devicepolicy.MixedDeviceOwnerTest#testKeyManagement
Bug: 138375478
Change-Id: Ibd6d840142cda6bc4fd5f28f4797867f91a5bb60