KeyChain tests: Add TEST_MAPPING
Add a test mapping for running KeyChain tests for any changes to the
KeyChain package.

Note that the Robolectric tests are not added yet because a test suite
cannot be specified for them.

The support service had to be built as an Android app (rather than
Android test) and specified as a dependency for the tests.

Bug: 138375478
Test: atest in packages/apps/KeyChain
Change-Id: Ia5d316f77791176ef63396a468b90491acc7a9eb