Fix KeyChain Robolectric tests
The KeyChain Robolectric tests were broken. Two fixes were necessary:
(1) Remove reference to the KeyChain AndroidManifest.xml file - it's not
needed and the file isn't available to the test for some reason.
(2) Make KeyChainService ask the injector for the calling app's UID.
This was broken by Change-Id: Iee37d48b82fda610f37912116ebe83357b46b53d
and is completely my fault for not putting the Robolectric tests in
presubmit earlier.

Bug: 72642900
Bug: 138375478
Test: atest KeyChainRoboTests
Change-Id: I525d26aad9b427bddfbb6c56eb39ff52e61af9bf