Inspect issuers when deciding on a certificate to display
To comply with RFC5246, KeyChain activity needs to inspect every
certificate in the certificate chain of a client certificate,
as the issuers list specified by the server could denote any
certificate in the chain.

Add a test to assert that specifying any issuer in the chain would cause
the certificate to be marked presentable, and modify KeyChain to load
and extract all issuers from the chain installed with the certificate.

To get the test running, it's fixed to be a proper instrumentation test
sharing the system UID so it can call code internal to the KeyChain
package.

Also fix broken service logging tests.

Bug: 142043186
Test: atest KeyChainTests
Change-Id: If03a71dcf0d8dc51f00f414c6aaad23b3ef2e952