Fix cert selection dialog background
Another theming fix: Change the background of the certificate selection
prompt such that it is transparent, showing the window behind it that
prompted the certificate selection process.

This is done by extending the Theme.DeviceDefault.Light theme, with a
transparent window color and no title bar.

Note that the "Transparent" theme in styles.xml was not used, so it was
re-purposed.

Bug: 143208078
Test: Manual, see description in change ID I76a35e33ab5e3228ff8c5f585edbd7c3f310a7c0
Change-Id: Ibb3829be8504475dd88f5c0fe8b04edb5f0109da