Make cert selection prompt respect dark mode
Use the DayNight theme in the KeyChain certificate selection prompt such
that it is displayed using a dark theme when dark mode is set.

Test: Manual, install TestDPC, set it as Device Owner, generate keys,
and observe that the prompt shown when choosing "Test signing/validation
with a key" respects dark mode.
Bug: 152854648

Merged-In: Ifef07bb5c2c8762ebe122d35883e1df4d059467b
Change-Id: I15006c6fabd497698e0390813ad6a1cdce5c2d09