Install CA certificates into KeyStore
When the user installs a CA certificate (via Settings), store that
certificate into the system KeyStore.

This is necessary for it to be available as a VPN trust anchor.
Note that this is a temporary fix and long-term the user should install
VPN trust anchor separately of CA certificates. Particularly in the
context of VPN, it's undesirable for the VPN provider's endpoint
authentication certificate to also be a root CA on the device, as all
traffic goes through the VPN.

Test: Manual, installed CA certificate via cert installer and ensured it's available in VPN settings.
Bug: 163413737
Change-Id: Ib6a308177b7ce6c55fb45af9809e0efc5c48a079
(cherry picked from commit a21a521ae9c1fbd099f0fd84df8776a0c1c020f8)